requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=py310h06a4308_0
altair=5.5.0=pypi_0
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-lru=2.0.4=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.9.0=pypi_0
bottleneck=1.3.7=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.5.0=pypi_0
certifi=2024.7.4=py310h06a4308_0
cffi=1.17.0=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=py310hdb19cb5_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.37=0
cuda-runtime=12.4.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
d2l=1.0.3=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h6a678d5_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py310h5eee18b_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
grpcio=1.62.2=py310h6a678d5_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py310h06a4308_0
imageio=2.35.1=pypi_0
importlib-metadata=8.4.0=pyha770c72_0
importlib_metadata=8.4.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.26.0=pyh707e725_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py310h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py310hff52083_0
jupyterlab=4.2.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kaggle=1.6.17=pyhd8ed1ab_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.11.0.15=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.37=0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libglib=2.78.4=hdc74915_0
libgomp=14.1.0=h77fa898_0
libgrpc=1.62.2=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hdb19cb5_3
libnpp=12.2.5.2=0
libnvfatbin=12.6.20=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.4.1=py310h06a4308_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
mysql=5.7.24=h721c034_2
narwhals=1.18.3=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py310h06a4308_0
notebook=7.2.2=pypi_0
notebook-shim=0.2.4=pypi_0
numexpr=2.8.7=py310h85018f9_0
numpy=1.23.5=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.3.1=hb9d3cd8_3
overrides=7.7.0=pypi_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.4.0=py310h5eee18b_0
pip=24.2=py310h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py310h06a4308_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pyha770c72_0
protobuf=4.25.3=py310h12ddb61_0
psutil=6.0.0=py310hc51659f_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=18.1.0=pypi_0
pycparser=2.22=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.10=py310h6a678d5_0
pyqt5-sip=12.13.0=py310h5eee18b_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.12=h955ad1f_0
python-dateutil=2.9.0post0=py310h06a4308_2
python-json-logger=2.0.7=pypi_0
python-slugify=5.0.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=2.4.0=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py310h06a4308_0
pyyaml=6.0.1=py310h5eee18b_0
pyzmq=25.1.2=py310h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
qtconsole=5.6.0=pypi_0
qtpy=2.4.1=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.20.0=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=72.1.0=py310h06a4308_0
sip=6.7.12=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.41.1=pypi_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tdqm=0.0.1=pypi_0
tenacity=9.0.0=pypi_0
tensorboard=2.17.0=py310h06a4308_0
tensorboard-data-server=0.7.0=py310h52d8a92_1
terminado=0.18.1=pypi_0
text-unidecode=1.3=pyhd3eb1b0_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.28=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
toml=0.10.2=pypi_0
tomli=2.0.1=py310h06a4308_0
torchaudio=2.4.0=py310_cu124
torchtriton=3.0.0=py310
torchvision=0.19.0=py310_cu124
tornado=6.4.1=py310hc51659f_0
tqdm=4.66.4=py310h2f386ee_0
traitlets=5.14.3=pyhd8ed1ab_0
types-python-dateutil=2.9.0.20240821=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py310h5eee18b_0
unidecode=1.2.0=pyhd3eb1b0_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=py310h06a4308_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=py310h06a4308_1
websocket-client=1.8.0=pypi_0
werkzeug=3.0.3=py310h06a4308_0
wheel=0.43.0=py310h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zipp=3.20.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
